+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 [02/25/2019]
+### Additions
+- Added [HSIMP](https://howsecureismypassword.net/)'s estimate on crack time.
+
+### Changes
+- Password box now shows dots instead of the password itself.
+
 ## 1.1.0 [02/11/2019]
 ### Additions
 - Actually made it use a real brute force method instead of randomly guessing.